requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asttokens==2.4.1
attrs==23.2.0
base58==1.0.3
bitarray==2.9.2
cached-property==1.5.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
cryptography==42.0.7
cytoolz==0.12.3
dataclassy==0.11.1
decorator==5.1.1
Deprecated==1.2.14
eip712==0.2.7
eth-account==0.11.2
eth-ape==0.7.23
eth-bloom==3.0.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-pydantic-types==0.1.0
eth-rlp==1.0.1
eth-tester==0.11.0b2
eth-typing==3.5.2
eth-utils==2.3.1
eth_abi==5.1.0
ethpm-types==0.6.11
evm-trace==0.1.5
evmchains==0.0.8
executing==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
hexbytes==0.3.1
idna==3.7
ijson==3.2.3
iniconfig==2.0.0
ipython==8.24.0
jedi==0.19.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lazyasd==0.1.4
lru-dict==1.2.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
morphys==1.0
msgspec==0.18.6
multidict==6.0.5
numpy==1.26.4
packaging==23.2
pandas==1.5.3
parsimonious==0.10.0
parso==0.8.4
pexpect==4.9.0
pluggy==1.5.0
prompt-toolkit==3.0.43
protobuf==5.27.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cid==0.3.0
py-ecc==7.0.1
py-evm==0.10.1b1
py-geth==4.4.0
py-multibase==1.0.3
py-multicodec==0.2.1
py-multihash==0.2.3
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
PyGithub==1.59.1
Pygments==2.18.0
PyJWT==2.8.0
PyNaCl==1.5.0
pytest==7.4.4
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.2
rich==13.7.1
rlp==4.0.1
rpds-py==0.18.1
ruff==0.4.5
safe-pysha3==1.0.4
semantic-version==2.10.0
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
stack-data==0.6.3
toolz==0.12.1
tqdm==4.66.4
traitlets==5.14.3
trie==3.0.1
typing_extensions==4.12.0
urllib3==2.2.1
varint==1.0.2
watchdog==3.0.0
wcwidth==0.2.13
web3==6.19.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
